refactor: nest chain adapters under hooks/chains
