+++ shared/src/hooks/chains/ethereum/usePetsContract.ts
