+++ shared/src/hooks/chains/ethereum/useVerifySignature.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { useApiClient } from '../../contexts/ApiClientContext';
+import { useApiClient } from '../../../contexts/ApiClientContext';
 
 interface VerifySignatureParams {
     address: string;
