+++ shared/src/hooks/chains/ethereum/useWatchPetsContract.ts
