+++ shared/src/hooks/chains/index.ts
@@ -1,0 +1,2 @@
+export * from './ethereum';
+export * from './solana';
