+++ shared/src/hooks/index.ts
@@ -1,3 +1,10 @@
-export * from './ethereum';
-export * from './solana';
-export * from './pets';
+export * from './chains';
+
+export { useActiveChain, type ActiveChain } from './useActiveChain';
+export { usePetList, type PetListResult } from './usePetList';
+export { useCreatePet, type CreatePetArgs, type PetMutationResult } from './useCreatePet';
+export { useLevelUpPet, type LevelUpPetArgs } from './useLevelUpPet';
+export { useRenamePet, type RenamePetArgs } from './useRenamePet';
+export { useBattlePets, type BattlePetsArgs } from './useBattlePets';
+export { useBreedPets, type BreedPetsArgs } from './useBreedPets';
+export { useTransferPet, type TransferPetArgs } from './useTransferPet';
