--- shared/src/hooks/pets/index.ts
@@ -1,14 +1,0 @@
-export { useActiveChain, type ActiveChain } from './useActiveChain';
-export { isActionSupported } from './featureSupport';
-export {
-    FeatureNotSupportedError,
-    NoActiveChainError,
-    type PetAction,
-} from './errors';
-export { usePetList, type PetListResult } from './usePetList';
-export { useCreatePet, type CreatePetArgs, type PetMutationResult } from './useCreatePet';
-export { useLevelUpPet, type LevelUpPetArgs } from './useLevelUpPet';
-export { useRenamePet, type RenamePetArgs } from './useRenamePet';
-export { useBattlePets, type BattlePetsArgs } from './useBattlePets';
-export { useBreedPets, type BreedPetsArgs } from './useBreedPets';
-export { useTransferPet, type TransferPetArgs } from './useTransferPet';
