+++ shared/src/hooks/useTransferPet.ts
@@ -1,21 +1,20 @@
 import { useState } from 'react';
-import { usePetsContract } from '../ethereum/usePetsContract';
-import { usePetActions } from '../solana/usePetActions';
-import { usePetsConfig } from '../../contexts/PetsConfigContext';
+import { usePetsContract } from './chains/ethereum/usePetsContract';
+import { usePetActions } from './chains/solana/usePetActions';
+import { usePetsConfig } from '../contexts/PetsConfigContext';
 import { useActiveChain } from './useActiveChain';
-import { isActionSupported } from './featureSupport';
-import { FeatureNotSupportedError, NoActiveChainError } from './errors';
+import { isActionSupported, FeatureNotSupportedError, NoActiveChainError } from '../utils/pets';
 import type { PetMutationResult } from './useCreatePet';
 
-export interface RenamePetArgs {
+export interface TransferPetArgs {
+    to: string;
     petId: string;
-    name: string;
 }
 
-export function useRenamePet(): PetMutationResult<RenamePetArgs> {
+export function useTransferPet(): PetMutationResult<TransferPetArgs> {
     const chain = useActiveChain();
     const { evm } = usePetsConfig();
-    const isSupported = isActionSupported(chain.kind === 'none' ? null : chain.kind, 'rename');
+    const isSupported = isActionSupported(chain.kind === 'none' ? null : chain.kind, 'transfer');
 
     const evmHook = usePetsContract({
         contractAddress: evm?.contractAddress,
@@ -26,18 +25,18 @@ export function useRenamePet(): PetMutationResult<RenamePetArgs> {
 
     const [localError, setLocalError] = useState<Error | null>(null);
 
-    const mutate = async (args: RenamePetArgs) => {
-        if (chain.kind === 'none') throw new NoActiveChainError('rename');
-        if (!isSupported) throw new FeatureNotSupportedError(chain.kind, 'rename');
+    const mutate = async (args: TransferPetArgs) => {
+        if (chain.kind === 'none') throw new NoActiveChainError('transfer');
+        if (!isSupported) throw new FeatureNotSupportedError(chain.kind, 'transfer');
         try {
             setLocalError(null);
             if (chain.kind === 'evm') {
-                evmHook.changeName(BigInt(args.petId), args.name);
+                evmHook.transferPet(args.to, BigInt(args.petId));
                 return;
             }
-            await solanaActions.renamePet.mutateAsync({
+            await solanaActions.transferPet.mutateAsync({
                 petId: Number(args.petId),
-                name: args.name,
+                to: args.to,
             });
         } catch (err) {
             const error = err instanceof Error ? err : new Error(String(err));
@@ -48,21 +47,22 @@ export function useRenamePet(): PetMutationResult<RenamePetArgs> {
 
     const reset = () => {
         setLocalError(null);
-        solanaActions.renamePet.reset();
+        solanaActions.transferPet.reset();
     };
 
     const isPending =
-        chain.kind === 'evm' ? evmHook.isPending : solanaActions.renamePet.isPending;
+        chain.kind === 'evm' ? evmHook.isPending : solanaActions.transferPet.isPending;
+
     const error =
         localError ??
         (chain.kind === 'evm'
             ? (evmHook.writeError as Error | null) ?? null
-            : (solanaActions.renamePet.error as Error | null) ?? null);
+            : (solanaActions.transferPet.error as Error | null) ?? null);
 
     const hash =
         chain.kind === 'evm'
             ? (evmHook.hash as string | undefined)
-            : (solanaActions.renamePet.data as string | undefined);
+            : (solanaActions.transferPet.data as string | undefined);
 
     return { isSupported, mutate, isPending, error, reset, hash };
 }
