+++ shared/src/index.ts
@@ -32,5 +32,5 @@ export {
     type PetsConfigProviderProps,
     type PetsEvmConfig,
 } from './contexts/PetsConfigContext';
-export type { Pet, PetChain } from './types/pet';
+export type { Pet, PetChain, PetAction } from './types/pet';
 export { queryClient } from './queryClient';
