+++ shared/src/utils/pets/index.ts
@@ -2,3 +2,5 @@ export { mapEvmPet, type EvmRawPet } from './mapEvmPet';
 export { mapSolanaPet, type SolanaPetAccountRow } from './mapSolanaPet';
 export { getRarityColor, getRarityName, isPetReadyAt } from './cosmetics';
 export { getReadyPets as getReadyPetsUnified, type ReadyPet } from './readyPets';
+export { isActionSupported } from './featureSupport';
+export { FeatureNotSupportedError, NoActiveChainError } from './errors';
